requirements.txt
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
deeptone==0.12.2
flake8==4.0.1
gevent==21.12.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.2.0
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.5
pathspec==0.9.0
platformdirs==2.5.1
PyAudio @ file:///C:/Users/Donald/Desktop/deeptone/LuVoTem/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
requests==2.27.1
scipy==1.7.3
SpeechRecognition==3.8.1
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.1.1
urllib3==1.26.9
websocket-client==1.3.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
